# docs/api/control-v1.openapi.yaml
openapi: 3.1.0
info:
  title: ya-router Control API
  version: 1.0.0
  description: 'Isolated management-plane contract for ya-routerd. The Control API
    does

    not accept YA_ROUTER_API_KEY and is not served from the OpenAI-compatible

    data-plane listener.

    '
servers:
- url: http://unix/control/v1
  description: Local Unix socket transport (socket path is deployment-specific)
- url: https://127.0.0.1:7443/control/v1
  description: Optional loopback TLS transport
paths:
...
  /providers:
    get:
      operationId: listControlProviders
      summary: List all compiled-in providers
      security: &id001
      - localUnixSocket: []
      - controlBearer: []
      - mutualTLS: []
      parameters:
      - $ref: '#/components/parameters/RequestID'
      - $ref: '#/components/parameters/ClientVersion'
      responses:
        '200':
          description: Successful redacted read
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderCollection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '426':
          $ref: '#/components/responses/UpgradeRequired'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /models:
    get:
      operationId: listControlModels
      summary: List model catalog freshness and availability
      security: *id001
      parameters:
      - $ref: '#/components/parameters/RequestID'
      - $ref: '#/components/parameters/ClientVersion'
      - name: refresh
        in: query
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Successful redacted read
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelCatalogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '426':
          $ref: '#/components/responses/UpgradeRequired'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    localUnixSocket:
      type: apiKey
      in: header
      name: X-Local-Unix-Socket-Identity
      description: 'Documentation marker only. Local identity is established by Unix
        socket

        ownership and mode; clients do not send this header.

        '
    controlBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-control-token
      description: 'Independent viewer/operator/admin control token. It must differ
        from

        YA_ROUTER_API_KEY and is allowed only for explicitly configured loopback

        TLS administration.

        '
    mutualTLS:
      type: mutualTLS
      description: 'Required for non-loopback remote administration. Verified URI
        SAN,

        full subject, or Common Name values map to viewer/operator/admin roles.

        '
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: Optional safe caller correlation ID. The service generates one
        otherwise.
    ClientVersion:
      name: X-YA-Client-Version
      in: header
      required: false
      schema:
        type: string
        examples:
        - 1.1.0
        - 1.0.0
      description: ya client semantic version used for N/N-1 negotiation.
...
  headers:
    RequestID:
      schema:
        type: string
      description: Correlation ID included in logs, audit, and typed errors.
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - retryable
      - request_id
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
        request_id:
          type: string
        details:
          type: object
          additionalProperties: true
    ProviderCollection:
      type: object
      required:
      - data
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProviderResource'
    ProviderResource:
      type: object
      required:
      - descriptor
      - enabled
      - effective_capabilities
      - health
      - generation
      properties:
        descriptor:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
        effective_capabilities:
          type: array
          items:
            type: string
        health:
          type: object
          additionalProperties: true
        generation:
          type: integer
          format: uint64
...
    ModelCatalogResponse:
      type: object
      required:
      - catalogs
      properties:
        catalogs:
          type: array
          items:
            $ref: '#/components/schemas/ProviderCatalog'
    ProviderCatalog:
      type: object
      required:
      - provider_id
      - enabled
      - available
      - models
      - stale
      properties:
        provider_id:
          type: string
        enabled:
          type: boolean
        available:
          type: boolean
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelResource'
        fetched_at:
          type: string
          format: date-time
        last_attempt_at:
          type: string
          format: date-time
        age_seconds:
          type: integer
          minimum: 0
        stale:
          type: boolean
        last_refresh_error:
          type: string
    ModelResource:
      type: object
      required:
      - id
      - upstream_id
      - provider_id
      - available
      properties:
        id:
          type: string
        upstream_id:
          type: string
        provider_id:
          type: string
        available:
          type: boolean
        owned_by:
          type: string
        supported_endpoints:
          type: array
          items:
            type: string
...
  responses:
    Unauthorized:
      description: Missing, invalid, or data-plane-only credential
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated identity lacks the required role
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Sanitized internal failure
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    UpgradeRequired:
      description: Client version is outside the supported N/N-1 window
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...